requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
colorama==0.4.6
Django==5.2.6
django-debug-toolbar==6.0.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
dotenv==0.9.9
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
